Delta Proportion Diverted

The proportion of delta inflow diverted from 1980-1999.

dlt_divers

Format

A 3 dimensional array: 12 by 20 by 2 [months, years, deltas]

[ , , 1] North Delta

[ , , 2] South Delta

Details

The North Delta is defined as the area west of and including the Sacramento River below Freeport to Chips Island.

The South Delta is defined as the area east of the Sacramento River below Freeport to Chips Island and the San Joaquin River below Vernalis.

For more details see:

use this link within R delta_flows

use this link if in a web browser