Load SIT Model Inputs for Years 1980-1999

returns list of data values used by model

load_baseline_data(species)

Arguments

species
Call function with one of the following options to load species specific model inputs: 'fall', 'winter', 'spring', 'steelhead'

Value

list of model inputs

Details

The list returned contains the following data:

p.tempMC20 = Proportion of days in month exceeding 20°C

p.diver = Proportion of flow diverted for each watershed

t.diver = Total flow diverted for each watershed

dlt.divers = Proportion of flow diverted in the North and South deltas

dlt.divers.tot = Total flow diverted in the North and South deltas

juv.tmp = Monthly mean temperature during juvenile outmigration in each watershed

juv.tmp.dlt = Monthly mean temperature during juvenile outmigration in the deltas

Dlt.inf = Total inflow in the deltas

DLThab = Habitat areas for the deltas

prop.Q.bypasses = Proportion of Sacramento River into bypasses

IChab.bypass = Instream habitat areas for each bypass

floodp.bypass = Floodplain habitat areas for each bypass

fp.weeks = Floodplain habitat duration inundated for each watershed

gate.top = Monthly record of bypasses overtopped

DegDay = Accumulated degree days for adults before spawning

retQ = Proportion of flows at tributary junction from natal watershed in October

upSacQ = Upper Sacramento flow

freeportQ = Flow at Georgiana slough

dlt.gates = Total and proportion number of days the cross channel gates are typically closed per month

egg.tmp.eff = Temperature effect on egg mortality

inps = Miscellaneous model inputs

Dlt.inp = Miscellaneous delta model inputs

prop.pulse = Proportion pulse flows

medQ = Monthly median flows

Fall Run Specific Outputs:

IChab.spawn = Spawning habitat areas for each watershed

IChab.fry = Fry instream habitat areas for each watershed

IChab.juv = Juvenile instream habitat areas for each watershed

floodP = Floodplain habitat areas for each watershed

Spring Run Specific Outputs:

IChab.spawn = Spawning habitat areas for each watershed

IChab.fry = Fry instream habitat areas for each watershed

IChab.juv = Juvenile instream habitat areas for each watershed

floodP = Floodplain habitat areas for each watershed

SR.pools = Holding Habitat

has.SR = Spring Run are present in watershed

Steelhead Specific Outputs:

IChab.spawn = Spawning habitat areas for each watershed

IChab.fry = Fry instream habitat areas for each watershed

IChab.juv = Juvenile instream habitat areas for each watershed

floodP = Floodplain habitat areas for each watershed

ST.pools = Holding Habitat

IChab.adult = Adult instream habitat areas for each watershed

meanQ = Monthly mean flow for each watershed

Winter Run Specific Outputs:

IChab.spawn = Spawning habitat areas for each Sacramento River segment

IChab.fry = Fry instream habitat areas for each Sacramento River segment

IChab.juv = Juvenile instream habitat areas for each Sacramento River segment

floodP = Floodplain habitat areas for each Sacramento River segment

Examples

# place these two lines within the SIT Salmon Population Model function to load the data
all_inputs <- cvpiaData::load_baseline_data('fall')
list2env(all_inputs, envir = environment())

#> <environment: 0x55828ca3c3d0>