Load SIT Model Inputs for Years 1980-1999

returns list of data values used by model

load_baseline_data(species)

Arguments

species
Call function with one of the following options to load species specific model inputs: 'fall', 'winter', 'spring', 'steelhead'

Value

list of model inputs

Details

The list returned contains the following data:

Miscellaneous model inputs

Proportion of flow diverted for each watershed

Total flow diverted for each watershed

Proportion of flow diverted in the North and South deltas

Total flow diverted in the North and South deltas

Monthly mean temperature during juvenile outmigration in each watershed

Monthly mean temperature during juvenile outmigration in the deltas

Total inflow in the deltas

Proportion of Sacramento River into Yolo Bypass

Proportion of Sacramento River into Sutter Bypass

Instream habitat areas for each watershed

Habitat areas for the deltas

Floodplain habitat areas for each watershed

Yearly record of bypasses overtopped in October of November

Accumulated degree days for adults before spawning

Proportion of flows at tributary junction from natal watershed in October

Upper Sacramento flow

Flow at Georgiana slough

Total and proportion number of days the cross channel gates are typically closed per month

Temperature effect on egg mortality

Examples

# place these two lines within the SIT Salmon Population Model function to load the data
all_inputs <- cvpiaData::load_baseline_data('fall')
list2env(all_inputs, envir = environment())

#> Error in list2env(all_inputs, envir = environment()): first argument must be a named list